requirements.txt
absl-py==2.1.0
asgiref==3.8.1
bleach==6.1.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
docopt==0.6.2
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
image==1.5.33
jsonpickle==3.2.2
kaggle==1.6.14
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
millify==0.1.1
mkl-service==2.4.0
munch==4.0.0
networkx==3.3
opencv-python==4.10.0.84
packaging==24.1
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
sacred==0.8.5
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
text-unidecode==1.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
webencodings==0.5.1
wrapt==1.16.0
